Move code to calculate uptime to DateTimeUtils
